+++ src/eio_writer.ml
@@ -1,27 +1,44 @@
-(* The record is not needed, it is just added to make sure we could add fields
-   in the implementation if we needed. *)
-type t = { buf_write : Eio.Buf_write.t }
+type t = Eio.Buf_write.t
 
-let buf_write t = t.buf_write
+let with_flow = Eio.Buf_write.with_flow
+let flush t = Eio.Buf_write.flush t
+let write_string t ?pos ?len str = Eio.Buf_write.string t ?off:pos ?len str
+let write_newline t = write_string t "\n"
 
-let with_flow ?initial_size flow f =
-  Eio.Buf_write.with_flow ?initial_size flow (fun buf_write -> f { buf_write })
+let write_line t str =
+  write_string t str;
+  write_newline t
 ;;
 
-let flush t = Eio.Buf_write.flush t.buf_write
-let print_string t str = Eio.Buf_write.string t.buf_write str
-let print_newline t = print_string t "\n"
+let writef t fmt = Stdlib.Format.kasprintf (fun str -> write_string t str) fmt
+let write_lines t lines = List.iter lines ~f:(fun line -> write_line t line)
 
-let print_endline t str =
-  print_string t str;
-  print_newline t
+let write_sexp ?(mach = false) t sexp =
+  write_line t (if mach then Sexp.to_string_mach sexp else Sexp.to_string_hum sexp)
 ;;
 
-let print_lines t lines = List.iter lines ~f:(fun line -> print_endline t line)
-
-let print_s ?(mach = false) t sexp =
-  print_endline t (if mach then Sexp.to_string_mach sexp else Sexp.to_string_hum sexp)
+let print_string ~env str =
+  with_flow (Eio.Stdenv.stdout env) (fun t -> write_string t str)
 ;;
 
-let printf t fmt = Printf.ksprintf (fun str -> print_string t str) fmt
-let aprintf t fmt = Stdlib.Format.kasprintf (fun str -> print_string t str) fmt
+let print_newline ~env = print_string ~env "\n"
+let print_endline ~env str = with_flow (Eio.Stdenv.stdout env) (fun t -> write_line t str)
+
+let print_lines ~env lines =
+  with_flow (Eio.Stdenv.stdout env) (fun t ->
+    List.iter lines ~f:(fun line -> write_line t line))
+;;
+
+let print_sexp ~env ?mach sexp =
+  with_flow (Eio.Stdenv.stdout env) (fun t -> write_sexp ?mach t sexp)
+;;
+
+let prerr_string ~env str =
+  with_flow (Eio.Stdenv.stderr env) (fun t -> write_string t str)
+;;
+
+let prerr_newline ~env = prerr_string ~env "\n"
+let prerr_endline ~env str = with_flow (Eio.Stdenv.stderr env) (fun t -> write_line t str)
+let printf ~env fmt = Printf.ksprintf (fun str -> print_string ~env str) fmt
+let aprintf ~env fmt = Stdlib.Format.kasprintf (fun str -> print_string ~env str) fmt
+let eprintf ~env fmt = Printf.ksprintf (fun str -> prerr_string ~env str) fmt