+++ CHANGES.md
@@ -1,3 +1,11 @@
+## 0.0.3 (2024-01-19)
+
+### Changed
+
+- Refactor the interface, use `Async.Writer` as model for the part operating on
+  `Eio_writer.t` and convert the `Stdio` part (`print_*`) into quick one-off
+  helpers. This is a breaking change.
+
 ## 0.0.2 (2024-01-18)
 
 ### Changed